feat(css-fundamentals): implement responsive header with React state
This change demonstrates using `useState` for managing the mobile navigation panel's open/close state and `useMediaQuery` for adapting the UI based on screen size. It also refactors the navigation into `NavHeader` and `Navbar` for improved component composition.
